+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - New outputs:
   - ODB: Exports the PCB in ODB++ format (KiCad 9) (#782)
+  - IPC2581: Exports the PCB in IPC-2581 (IPC-DPMX) format (KiCad 9)
 - CLI:
   - `--variant` to specify which variants will be generated (See #737)
   - `--defs-from-project` allows using KiCad variables for the preprocessor